Remove some dead code
- Remove unused methods on Glyphs
- Made linear/log axis instances of Axis, not instances of subclasses
- Move py2/3 check in utils to compatibility, add test_compatibility
