+++ datashader/compatibility.py
@@ -10,5 +10,12 @@ if PY3:
             return func(*args, **kwargs)
         else:
             return func(*args)
+
+    def _exec(codestr, glbls):
+        exec(codestr, glbls)
 else:
     apply = apply
+    eval(compile("""
+def _exec(codestr, glbls):
+    exec codestr in glbls
+""", "<_exec>", "exec"))
