+++ datashader/tests/test_utils.py
@@ -1,24 +1,6 @@
 from datashape import dshape
 
-from datashader.utils import _exec, Dispatcher, isreal
-
-from pytest import raises
-
-
-def test__exec():
-    c = "def foo(a):\n    return bar(a) + 1"
-    namespace = {'bar': lambda a: a + 1}
-    # Define a different local ``bar`` to ensure that names are pulled from
-    # namespace, not locals
-    bar = lambda a: a - 1
-    _exec(c, namespace)
-    foo = namespace['foo']
-    assert foo(1) == 3
-    namespace = {}
-    _exec(c, namespace)
-    foo = namespace['foo']
-    with raises(NameError):
-        foo(1)
+from datashader.utils import Dispatcher, isreal
 
 
 def test_Dispatcher():
